requirements.txt
numpy==1.24.2
pandas==1.5.3
scikit-learn==1.2.2
python-dotenv==1.0.0
mysql-client==0.0.1
mysql-connector-python==8.0.32
mysqlclient==2.1.1
pyodbc==4.0.35
ipykernel==6.19.2
fastapi==0.95.0
uvicorn==0.21.1
streamlit==1.21.0
apache-airflow==2.6.1
great-expectations==0.16.16